begin basic param-gui test

+++ mne_pipeline_hd/development/dev_utils.py
@@ -94,7 +94,7 @@ class Widget(QWidget):
             btn = QPushButton(i)
             btn.setIcon(self.style().standardIcon(getattr(QStyle, i)))
 
-            layout.addWidget(btn, count / colSize, count % colSize)
+            layout.addWidget(btn, int(count / colSize), int(count % colSize))
             count += 1
 
         self.setLayout(layout)
